requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.6.1
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
replicate==0.22.0
gunicorn==21.2.0